+++ Gromit-client-iOS/Views/ChallengeListView.swift
@@ -10,6 +10,7 @@ import SwiftUI
 struct ChallengeListView: View {
     var body: some View {
         VStack {
+        
             ChallengeButtons()
             
             ChallengeCell()
@@ -28,49 +29,28 @@ struct ChallengeButtons: View {
     
     var body: some View {
         HStack {
-            Spacer()
-            Button("참여 챌린지") {
-                showParticipating.toggle()
+            Group{
+                Spacer()
+                Button("참여 챌린지") {
+                    showParticipating.toggle()
+                }
+                .fullScreenCover(isPresented: $showParticipating) {
+                    CreationView()
+                }
+                .buttonStyle(.bordered)
+                .cornerRadius(20)
             }
-            .fullScreenCover(isPresented: $showParticipating) {
-                ParticipatingListView()
-            }
-            .buttonStyle(.bordered)
-            .cornerRadius(20)
             
-            Button("+") {
-                
-            }
-            .buttonStyle(.bordered)
-            .cornerRadius(20)
-        }
-        .padding(EdgeInsets(top: 0, leading: 0, bottom: 0, trailing: 10))
-    }
-}
-
-struct ChallengeCell: View {
-    var body: some View {
-        List(0..<20) { item in
-            VStack(alignment: .leading, spacing: 5) {
-                Text("20자이내로제목을작성해주세요감사합니다")
-                    .fontWeight(.semibold)
-                    .lineLimit(1)
-                    .padding(EdgeInsets(top: 0, leading: 0, bottom: 10, trailing: 0))
-                HStack {
-                    Spacer()
-                    Text("23/01/03")
+            Group{
+                Button("+") {
+                    showParticipating.toggle()
                 }
-                HStack {
-                    Text("목표 커밋: 100")
-                    Spacer()
-                    Text("1/6")
+                .fullScreenCover(isPresented: $showParticipating) {
+                    CreationView()
                 }
+                .buttonStyle(.bordered)
+                .cornerRadius(20)
             }
-            .padding(EdgeInsets(top: 20, leading: 30, bottom: 20, trailing: 30))
-            .background(Color("yellow500"))
-            .cornerRadius(20)
-            .shadow(color: Color("gray500"), radius: 5, y: 5)
-        }
-        .listStyle(PlainListStyle())
-    }
+            .padding(EdgeInsets(top: 0, leading: 0, bottom: 0, trailing: 10))
+        }}
 }
